interactive_marker: add detection for relative objects when in camera image
