+++ humanoid_league_interactive_marker/launch/interactive_marker.launch
@@ -1,0 +1,9 @@
+<launch>
+    <rosparam file="$(find humanoid_league_interactive_marker)/config/camera_info.yaml" command="load"/>
+
+    <node name="rviz" pkg="rviz" type="rviz" output="screen" required="true"
+          args="-d $(find humanoid_league_interactive_marker)/config/interactive_marker.rviz">
+    </node>
+    <node name="interactive_marker" pkg="humanoid_league_interactive_marker" type="interactive_marker.py" output="screen"/>
+
+</launch>
