--- humanoid_league_interactive_marker/launch/rviz_behavior_test.launch
@@ -1,25 +1,0 @@
-<launch>
-    <param name="visualization_active" value="true" />
-    <include file="$(find bitbots_bringup)/launch/load_robot_description.launch">
-        <arg name="wolfgang" value="true"/>
-    </include>
-
-    <include file="$(find bitbots_move_base)/launch/pathfinding_move_base.launch" />
-
-    <rosparam file="$(find bitbots_quintic_walk)/config/walking_wolfgang_viz.yaml" command="load"/>
-    <node name="walking" pkg="bitbots_quintic_walk" type="WalkNode" respawn="true" output="screen"/>
-
-    <include file="$(find bitbots_animation_server)/launch/animation.launch" />
-    <include file="$(find bitbots_hcm)/launch/hcm.launch">
-        <arg name="sim" value="true"/>
-    </include>
-
-    <node name="rviz" pkg="rviz" type="rviz" output="screen" required="true"
-          args="-d $(find humanoid_league_interactive_marker)/config/behavior_test.rviz">
-    </node>
-
-    <node name="ball_marker" pkg="humanoid_league_interactive_marker" type="rviz_behavior_test.py"/>
-    <node name="motor_goals_viz_helper" pkg="bitbots_bringup" type="motor_goals_viz_helper.py"/>
-    <!-- TODO: replace this with the launch file when one exists -->
-    <node name="dynamic_kick" pkg="bitbots_dynamic_kick" type="KickNode" output="screen" />
-</launch>
